% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boosh_phonx.r
\name{choozeRound}
\alias{choozeRound}
\title{a quik-round rounder}
\usage{
choozeRound(x, d = 2)
}
\arguments{
\item{x}{a num or int (atomic or vector) or other obj}

\item{d}{a num, specifies how many diggies to round x to. Defaults to 2.}
}
\description{
tha choozeRound phonk takes a num + rounds it how u want + gives it baqq <3
}
\examples{
choozeRound()
}
\keyword{arithmetic,}
\keyword{numbers}
\keyword{round,}

